renamed qr package to qrcode and moved into connection settings
