+++ app/src/main/java/com/babylon/wallet/android/presentation/settings/addconnection/SettingsConnectionScreen.kt
@@ -39,7 +39,7 @@ import com.babylon.wallet.android.designsystem.composable.RadixTextField
 import com.babylon.wallet.android.designsystem.theme.RadixTheme
 import com.babylon.wallet.android.designsystem.theme.RadixWalletTheme
 import com.babylon.wallet.android.presentation.common.FullscreenCircularProgressContent
-import com.babylon.wallet.android.presentation.qr.CameraPreview
+import com.babylon.wallet.android.presentation.settings.addconnection.qrcode.CameraPreview
 import com.babylon.wallet.android.presentation.ui.composables.RadixCenteredTopAppBar
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
 import com.google.accompanist.permissions.PermissionState
